requirements.txt
google-cloud-bigquery
flask
pandas
numpy
matplotlib
seaborn
scikit-learn
tensorflow
pytest
statsmodels
db-dtypes
joblib
